rename `to_geocode` to `geocode`
