+++ dtoolkit/geoaccessor/dataframe/__init__.py
@@ -1,4 +1,4 @@
 from dtoolkit.geoaccessor.dataframe.from_wkt import from_wkt  # noqa
 from dtoolkit.geoaccessor.dataframe.from_xy import from_xy  # noqa
-from dtoolkit.geoaccessor.dataframe.to_geocode import to_geocode  # noqa
+from dtoolkit.geoaccessor.dataframe.geocode import geocode  # noqa
 from dtoolkit.geoaccessor.dataframe.to_geoframe import to_geoframe  # noqa
