+++ dtoolkit/geoaccessor/series/__init__.py
@@ -1,1 +1,1 @@
-from dtoolkit.geoaccessor.series.to_geocode import to_geocode  # noqa
+from dtoolkit.geoaccessor.series.geocode import geocode  # noqa
